KA-160 refactor test and folder structure
