+++ onboarding-task/src/utils/isTextInputValid.js
@@ -1,3 +1,1 @@
-export function isTextInputValid(input) {
-  return (input && input.trim().length > 0) === true;
-}
+export const isTextInputValid = (input) => input && input.trim().length > 0;
